third_party/blink/web_tests/virtual/partitioned-cookies/external/wpt/html/cross-origin-embedder-policy/anonymous-iframe/cookie.tentative.https.window-expected.txt

This is a testharness.js-based test.
PASS Setup
FAIL Anonymous same-origin iframe is loaded without credentials assert_equals: expected (undefined) undefined but got (string) "same_origin"
FAIL Anonymous cross-origin iframe is loaded without credentials assert_equals: expected (undefined) undefined but got (string) "cross_origin"
PASS same_origin anonymous iframe can't send same_origin credentials
PASS same_origin anonymous iframe can't send cross_origin credentials
PASS cross_origin anonymous iframe can't send cross_origin credentials
PASS cross_origin anonymous iframe can't send same_origin credentials
PASS same_origin anonymous iframe can't send same_origin credentials on child iframe
PASS same_origin anonymous iframe can't send cross_origin credentials on child iframe
PASS cross_origin anonymous iframe can't send cross_origin credentials on child iframe
PASS cross_origin anonymous iframe can't send same_origin credentials on child iframe
Harness: the test ran to completion.
